refactor: LikeService 코드수정

+++ src/main/java/com/example/inflearn/service/like/LikeService.java
@@ -27,16 +27,20 @@ public class LikeService {
     public void saveLike(long memberId, long postId) {
         Member member = memberRepository.findById(memberId).orElseThrow(DoesNotExistMemberException::new);
         Post post = postRepository.findById(postId).orElseThrow(DoesNotExistPostException::new);
-        likeRepository.findByMemberAndPost(member, post).ifPresent(like -> {
+        Like byMemberAndPost = likeRepository.findByMemberAndPost(member, post);
+        if (byMemberAndPost != null) {
             throw new AlreadyLikeException();
-        });
+        }
         likeRepository.save(Like.create(member, post));
     }
 
     public void unLike(long memberId, long postId) {
         Member member = memberRepository.findById(memberId).orElseThrow(DoesNotExistMemberException::new);
         Post post = postRepository.findById(postId).orElseThrow(DoesNotExistPostException::new);
-        Like like = likeRepository.findByMemberAndPost(member, post).orElseThrow(DoesNotLikeException::new);
+        Like like = likeRepository.findByMemberAndPost(member, post);
+        if (like == null) {
+            throw new DoesNotLikeException();
+        }
         likeRepository.delete(like);
     }
 }
